Move threshold application and conversion to detector_task, drop DetectionResult, reuse type aliases
